Move API tests to IngestGranuleSuccessSpec
