--- example/spec/testAPI/testAPI.input.payload.json
@@ -1,32 +1,0 @@
-{
-    "granules": [
-      {
-        "granuleId": "MOD09GQ.A2016358.h13v04.006.2016360104606",
-        "dataType": "MOD09GQ",
-        "version": "006",
-        "files": [
-          {
-            "path": "cumulus-test-data/pdrs",
-            "fileSize": 17865615,
-            "name": "MOD09GQ.A2016358.h13v04.006.2016360104606.hdf"
-          },
-          {
-            "path": "cumulus-test-data/pdrs",
-            "fileSize": 44118,
-            "name": "MOD09GQ.A2016358.h13v04.006.2016360104606.hdf.met"
-          },
-          {
-            "path": "cumulus-test-data/pdrs",
-            "fileSize": 44118,
-            "name": "MOD09GQ.A2016358.h13v04.006.2016360104606_ndvi.jpg"
-          }
-        ]
-      }
-    ],
-    "pdr": {
-      "path": "cumulus-test-data/pdrs",
-      "size": 596,
-      "name": "MOD09GQ_1granule_v3.PDR",
-      "time": 1520873050000
-    }
-}
